Re-fatoring of io
Refactoring in progress of the io module. Changes include:

1) Both transient and persistent classes defining the event model are
now in reco/event_model.
2) new file table_io, defining a general make_table function.
3) old dst_io has been split into kdst_io (Dorothea kdst io) and
dst_io (Penthesilea DST io).

The corresponding modified imports have been propagated to relevant
files.
