+++ invisible_cities/reco/nh5.py
@@ -159,12 +159,14 @@ class Tfactors(tb.IsDescription):
 
 
 class HitsTable(tb.IsDescription):
-    event = tb.  Int32Col(pos=0)
-    time  = tb.Float64Col(pos=1)
-    npeak = tb. UInt16Col(pos=2)
-    nsipm = tb. UInt16Col(pos=3)
-    X     = tb.Float64Col(pos=4)
-    Y     = tb.Float64Col(pos=5)
-    Z     = tb.Float64Col(pos=6)
-    Q     = tb.Float64Col(pos=7)
-    E     = tb.Float64Col(pos=8)
+    # event = tb.  Int32Col(pos=0)
+    # time  = tb.Float64Col(pos=1)
+    npeak    = tb. UInt16Col(pos=0)
+    nsipm    = tb. UInt16Col(pos=1)
+    X        = tb.Float64Col(pos=2)
+    Y        = tb.Float64Col(pos=3)
+    Xrms     = tb.Float64Col(pos=4)
+    Yrms     = tb.Float64Col(pos=5)
+    Z        = tb.Float64Col(pos=6)
+    Q        = tb.Float64Col(pos=7)
+    E        = tb.Float64Col(pos=8)
